Temporarily ban repeated failed authorizations to /claim-key
I had previously intended for this to be done externally with a WAF or
something, but I'm now starting to think that this is such a critical
protection - and AWS WAF is more limited than I realized - that we
should do it internally to simplify deployment.

I considered adding an extra proxy component to sit in front of this
one, in order to keep IP addresses (PII) out of the same database as the
diagnosis keys, but:

* IP addresses are deleted before adding diagnosis keys to the database,
  so the database should never concurrently contain both PII and any keys
  relating to that I'd P.
* IP addresses are only stored in the first place after failed key claim
  attempts, and even then, only for the ban duration (1 hour currently)
  if no successful authentication occurs (otherwise deleted sooner).

This ultimately has to get stored somewhere in order to implement this
protection and it doesn't appear to me that there's an additional risk
in having it live here; quite the opposite, the simplicity of having
only one component makes me more confident that we have implemented this
correctly.

This commit is currently in a "well it compiles; maybe it works" state.
Comprehensive testing required before merge.