Perist keyclaim token that generated keys:
Keyclaim tokens will inevitably be compromised: It's probably necessary
to have this feature to be able to purge spammed tokens from compromised
sources.

Note what isn't implemented here: This is kind of calling out to be
turned into a richer model with authentication for keyclaim clients
stored in a database table, rather than simply in an environment
variable, and linking each key via a foreign key to those records,
rather than embedding the authentication token directly.

I *think* this is not necessary: the additional complexity would be an
operational burden, and though this means more data storage per key, I
think this is a tolerable tradeoff.