Action for Creating an invalidation on config.go
This will naively invalidated the exposure configuration path whenever
the pkg/server/config.go file get's changed. Though considering this
file was modified in pretty much all but one of the commits of this
files history I think it's a safe bet.

This should prevent someone from accidentally invalidating /* when the
configuration changes. This can also be modified when we put the
configurations in an s3 bucket to invalidate the CDN in front of
those.