Fix nil dereference if timestamp is not specified
The protobuf definition for Upload shows timestamp as an optional field:

	optional google.protobuf.Timestamp timestamp = 1;

As such, the server will demarshal Uploads that lack a timestamp.
However, the server will later dereference the timestamp to access the
seconds count, reliably panicking the server and unexpectedly breaking
the connection to the client.

The panic is caught by safely.Recover, so the harm is minimized.
Further, executing this code path requires a COVID-19 diagnosis.
Nevertheless, the behaviour is still incorrect and should be hardened.

By explicitly checking if timestamp is non-nil, the panic is fixed and
INVALID_TIMESTAMP is returned as expected.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>